pyproject.toml
[project]
name = "switchmap-py"
version = "0.1.0"
description = "Python reimplementation of dotysan/switchmap"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "OpenAI Codex"}]
keywords = ["snmp", "network", "switchmap"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.2.1",
  "pyyaml>=6.0.1",
  "jinja2>=3.1.3",
]

[project.optional-dependencies]
search = ["fastapi>=0.111.0", "uvicorn>=0.30.0"]
snmp = ["pysnmp>=4.4.12"]

[project.scripts]
switchmap = "switchmap_py.cli:app"

[tool.pytest.ini_options]
pythonpath = ["."]

testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69.5.1", "wheel"]
build-backend = "setuptools.build_meta"
